Cleanup API endpoints
This PR cleans the API used by the dashboard:
- Moves all the endpoints to the `/api/v1.py` to avoid confusion
- Removes the unused endpoints
- Moves all the models and processing to `/api/v1_models.py` and `/api/v1_processing.py`
